+++ composer.json
@@ -26,6 +26,7 @@
         "monolog/monolog": "~1.11",
         "mtdowling/cron-expression": "~1.0",
         "nesbot/carbon": "~1.19",
+        "paragonie/random_compat": "^1.0.4",
         "psy/psysh": "~0.5.1",
         "swiftmailer/swiftmailer": "~5.1",
         "symfony/console": "2.7.*",
