+++ src/Illuminate/Auth/Access/Gate.php
@@ -2,12 +2,15 @@
 
 namespace Illuminate\Auth\Access;
 
+use Illuminate\Support\Str;
 use InvalidArgumentException;
 use Illuminate\Contracts\Container\Container;
 use Illuminate\Contracts\Auth\Access\Gate as GateContract;
 
 class Gate implements GateContract
 {
+    use HandlesAuthorization;
+
     /**
      * The container instance.
      *
@@ -44,6 +47,13 @@ class Gate implements GateContract
     protected $beforeCallbacks = [];
 
     /**
+     * All of the registered after callbacks.
+     *
+     * @var array
+     */
+    protected $afterCallbacks = [];
+
+    /**
      * Create a new gate instance.
      *
      * @param  \Illuminate\Contracts\Container\Container  $container
@@ -51,14 +61,16 @@ class Gate implements GateContract
      * @param  array  $abilities
      * @param  array  $policies
      * @param  array  $beforeCallbacks
+     * @param  array  $afterCallbacks
      * @return void
      */
-    public function __construct(Container $container, callable $userResolver, array $abilities = [], array $policies = [], array $beforeCallbacks = [])
+    public function __construct(Container $container, callable $userResolver, array $abilities = [], array $policies = [], array $beforeCallbacks = [], array $afterCallbacks = [])
     {
         $this->policies = $policies;
         $this->container = $container;
         $this->abilities = $abilities;
         $this->userResolver = $userResolver;
+        $this->afterCallbacks = $afterCallbacks;
         $this->beforeCallbacks = $beforeCallbacks;
     }
 
@@ -138,6 +150,19 @@ class Gate implements GateContract
     }
 
     /**
+     * Register a callback to run after all Gate checks.
+     *
+     * @param  callable  $callback
+     * @return $this
+     */
+    public function after(callable $callback)
+    {
+        $this->afterCallbacks[] = $callback;
+
+        return $this;
+    }
+
+    /**
      * Determine if the given ability should be granted for the current user.
      *
      * @param  string  $ability
@@ -170,21 +195,78 @@ class Gate implements GateContract
      */
     public function check($ability, $arguments = [])
     {
+        try {
+            $result = $this->raw($ability, $arguments);
+        } catch (UnauthorizedException $e) {
+            return false;
+        }
+
+        return (bool) $result;
+    }
+
+    /**
+     * Determine if the given ability should be granted for the current user.
+     *
+     * @param  string  $ability
+     * @param  array|mixed  $arguments
+     * @return \Illuminate\Auth\Access\Response
+     *
+     * @throws \Illuminate\Auth\Access\UnauthorizedException
+     */
+    public function authorize($ability, $arguments = [])
+    {
+        $result = $this->raw($ability, $arguments);
+
+        if ($result instanceof Response) {
+            return $result;
+        }
+
+        return $result ? $this->allow() : $this->deny();
+    }
+
+    /**
+     * Get the raw result for the given ability for the current user.
+     *
+     * @param  string  $ability
+     * @param  array|mixed  $arguments
+     * @return mixed
+     */
+    protected function raw($ability, $arguments = [])
+    {
         if (! $user = $this->resolveUser()) {
             return false;
         }
 
         $arguments = is_array($arguments) ? $arguments : [$arguments];
 
-        if (! is_null($result = $this->callBeforeCallbacks($user, $ability, $arguments))) {
-            return $result;
+        if (is_null($result = $this->callBeforeCallbacks($user, $ability, $arguments))) {
+            $result = $this->callAuthCallback($user, $ability, $arguments);
         }
 
+        $this->callAfterCallbacks(
+            $user, $ability, $arguments, $result
+        );
+
+        return $result;
+    }
+
+    /**
+     * Resolve and call the appropriate authorization callback.
+     *
+     * @param  \Illuminate\Contracts\Auth\Authenticatable  $user
+     * @param  string  $ability
+     * @param  array  $arguments
+     * @return bool
+     */
+    protected function callAuthCallback($user, $ability, array $arguments)
+    {
         $callback = $this->resolveAuthCallback(
             $user, $ability, $arguments
         );
 
-        return call_user_func_array($callback, array_merge([$user], $arguments));
+        return call_user_func_array(
+            $callback, array_merge([$user], $arguments)
+        );
     }
 
     /**
@@ -203,6 +285,24 @@ class Gate implements GateContract
             if (! is_null($result = call_user_func_array($before, $arguments))) {
                 return $result;
             }
+        }
+    }
+
+    /**
+     * Call all of the after callbacks with check result.
+     *
+     * @param  \Illuminate\Contracts\Auth\Authenticatable  $user
+     * @param  string  $ability
+     * @param  array  $arguments
+     * @param  bool  $result
+     * @return void
+     */
+    protected function callAfterCallbacks($user, $ability, array $arguments, $result)
+    {
+        $arguments = array_merge([$user, $ability, $result], $arguments);
+
+        foreach ($this->afterCallbacks as $after) {
+            call_user_func_array($after, $arguments);
         }
     }
 
@@ -275,6 +375,10 @@ class Gate implements GateContract
                 }
             }
 
+            if (strpos($ability, '-') !== false) {
+                $ability = Str::camel($ability);
+            }
+
             if (! is_callable([$instance, $ability])) {
                 return false;
             }
@@ -326,7 +430,8 @@ class Gate implements GateContract
     public function forUser($user)
     {
         return new static(
-            $this->container, function () use ($user) { return $user; }, $this->abilities, $this->policies, $this->beforeCallbacks
+            $this->container, function () use ($user) { return $user; }, $this->abilities,
+            $this->policies, $this->beforeCallbacks, $this->afterCallbacks
         );
     }
