+++ src/Illuminate/Auth/Access/HandlesAuthorization.php
@@ -1,0 +1,29 @@
+<?php
+
+namespace Illuminate\Auth\Access;
+
+trait HandlesAuthorization
+{
+    /**
+     * Create a new access response.
+     *
+     * @return \Illuminate\Auth\Access\Response
+     */
+    protected function allow($message = null)
+    {
+        return new Response($message);
+    }
+
+    /**
+     * Throws an unauthorized exception.
+     *
+     * @param  string  $message
+     * @return void
+     *
+     * @throws \Illuminate\Auth\Access\UnauthorizedException
+     */
+    protected function deny($message = 'This action is unauthorized.')
+    {
+        throw new UnauthorizedException($message);
+    }
+}
