+++ src/Illuminate/Auth/Access/UnauthorizedException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Illuminate\Auth\Access;
+
+use Exception;
+
+class UnauthorizedException extends Exception
+{
+    //
+}
