+++ src/Illuminate/Auth/Guard.php
@@ -133,9 +133,9 @@ class Guard implements GuardContract
             return;
         }
 
-        // If we have already retrieved the user for the current request we can just
-        // return it back immediately. We do not want to pull the user data every
-        // request into the method because that would tremendously slow an app.
+        // If we've already retrieved the user for the current request we can just
+        // return it back immediately. We do not want to fetch the user data on
+        // every call to this method because that would be tremendously slow.
         if (! is_null($this->user)) {
             return $this->user;
         }
