+++ src/Illuminate/Database/Connectors/ConnectionFactory.php
@@ -98,6 +98,11 @@ class ConnectionFactory
     {
         $readConfig = $this->getReadWriteConfig($config, 'read');
 
+        if (isset($readConfig['host']) && is_array($readConfig['host'])) {
+            $readConfig['host'] = count($readConfig['host']) > 1 ?
+                $readConfig['host'][array_rand($readConfig['host'])] : $readConfig['host'][0];
+        }
+
         return $this->mergeReadWriteConfig($config, $readConfig);
     }
