+++ src/Illuminate/Database/Eloquent/FactoryBuilder.php
@@ -124,7 +124,7 @@ class FactoryBuilder
     {
         return Model::unguarded(function () use ($attributes) {
             if (! isset($this->definitions[$this->class][$this->name])) {
-                throw new InvalidArgumentException("Unable to locate factory with name [{$this->name}].");
+                throw new InvalidArgumentException("Unable to locate factory with name [{$this->name}] [{$this->class}].");
             }
 
             $definition = call_user_func($this->definitions[$this->class][$this->name], $this->faker, $attributes);
