+++ src/Illuminate/Database/Query/Grammars/Grammar.php
@@ -131,8 +131,6 @@ class Grammar extends BaseGrammar
     {
         $sql = [];
 
-        $query->setBindings([], 'join');
-
         foreach ($joins as $join) {
             $table = $this->wrapTable($join->table);
 
@@ -143,10 +141,6 @@ class Grammar extends BaseGrammar
 
             foreach ($join->clauses as $clause) {
                 $clauses[] = $this->compileJoinConstraint($clause);
-            }
-
-            foreach ($join->bindings as $binding) {
-                $query->addBinding($binding, 'join');
             }
 
             // Once we have constructed the clauses, we'll need to take the boolean connector
