+++ src/Illuminate/Foundation/Auth/Access/AuthorizesRequests.php
@@ -3,6 +3,7 @@
 namespace Illuminate\Foundation\Auth\Access;
 
 use Illuminate\Contracts\Auth\Access\Gate;
+use Illuminate\Auth\Access\UnauthorizedException;
 use Symfony\Component\HttpKernel\Exception\HttpException;
 
 trait AuthorizesRequests
@@ -12,7 +13,7 @@ trait AuthorizesRequests
      *
      * @param  mixed  $ability
      * @param  mixed|array  $arguments
-     * @return void
+     * @return \Illuminate\Auth\Access\Response
      *
      * @throws \Symfony\Component\HttpKernel\Exception\HttpException
      */
@@ -20,9 +21,7 @@ trait AuthorizesRequests
     {
         list($ability, $arguments) = $this->parseAbilityAndArguments($ability, $arguments);
 
-        if (! app(Gate::class)->check($ability, $arguments)) {
-            throw $this->createGateUnauthorizedException($ability, $arguments);
-        }
+        return $this->authorizeAtGate(app(Gate::class), $ability, $arguments);
     }
 
     /**
@@ -31,7 +30,7 @@ trait AuthorizesRequests
      * @param  \Illuminate\Contracts\Auth\Authenticatable|mixed  $user
      * @param  mixed  $ability
      * @param  mixed|array  $arguments
-     * @return void
+     * @return \Illuminate\Auth\Access\Response
      *
      * @throws \Symfony\Component\HttpKernel\Exception\HttpException
      */
@@ -39,10 +38,29 @@ trait AuthorizesRequests
     {
         list($ability, $arguments) = $this->parseAbilityAndArguments($ability, $arguments);
 
-        $result = app(Gate::class)->forUser($user)->check($ability, $arguments);
+        $gate = app(Gate::class)->forUser($user);
 
-        if (! $result) {
-            throw $this->createGateUnauthorizedException($ability, $arguments);
+        return $this->authorizeAtGate($gate, $ability, $arguments);
+    }
+
+    /**
+     * Authorize the request at the given gate.
+     *
+     * @param  \Illuminate\Contracts\Auth\Access\Gate  $gate
+     * @param  mixed  $ability
+     * @param  mixed|array  $arguments
+     * @return \Illuminate\Auth\Access\Response
+     *
+     * @throws \Symfony\Component\HttpKernel\Exception\HttpException
+     */
+    public function authorizeAtGate(Gate $gate, $ability, $arguments)
+    {
+        try {
+            return $gate->authorize($ability, $arguments);
+        } catch (UnauthorizedException $e) {
+            throw $this->createGateUnauthorizedException(
+                $ability, $arguments, $e->getMessage(), $e
+            );
         }
     }
 
@@ -66,11 +84,13 @@ trait AuthorizesRequests
      * Throw an unauthorized exception based on gate results.
      *
      * @param  string  $ability
-     * @param  array  $arguments
+     * @param  mixed|array  $arguments
+     * @param  string  $message
+     * @param  \Exception  $previousException
      * @return \Symfony\Component\HttpKernel\Exception\HttpException
      */
-    protected function createGateUnauthorizedException($ability, $arguments)
+    protected function createGateUnauthorizedException($ability, $arguments, $message = 'This action is unauthorized.', $previousException = null)
     {
-        return new HttpException(403, 'This action is unauthorized.');
+        return new HttpException(403, $message, $previousException);
     }
 }
