+++ src/Illuminate/Foundation/Console/Kernel.php
@@ -35,6 +35,13 @@ class Kernel implements KernelContract
     protected $artisan;
 
     /**
+     * The Artisan commands provided by the application.
+     *
+     * @var array
+     */
+    protected $commands = [];
+
+    /**
      * The bootstrap classes for the application.
      *
      * @var array
