+++ src/Illuminate/Foundation/Console/ServeCommand.php
@@ -28,6 +28,8 @@ class ServeCommand extends Command
      * Execute the console command.
      *
      * @return void
+     *
+     * @throws \Exception
      */
     public function fire()
     {
