+++ src/Illuminate/Routing/Route.php
@@ -160,6 +160,8 @@ class Route
      *
      * @param  \Illuminate\Http\Request  $request
      * @return mixed
+     *
+     * @throws \Symfony\Component\HttpKernel\Exception\NotFoundHttpException
      */
     protected function runController(Request $request)
     {
@@ -454,7 +456,7 @@ class Route
      *
      * @return array
      *
-     * @throws LogicException
+     * @throws \LogicException
      */
     public function parameters()
     {
@@ -921,8 +923,7 @@ class Route
     /**
      * Add or change the route name.
      *
-     * @param $name
-     *
+     * @param  string  $name
      * @return $this
      */
     public function name($name)
@@ -993,7 +994,7 @@ class Route
      *
      * @return void
      *
-     * @throws LogicException
+     * @throws \LogicException
      */
     public function prepareForSerialization()
     {
