+++ src/Illuminate/Routing/Router.php
@@ -723,7 +723,7 @@ class Router implements RegistrarContract
     /**
      * Resolve the middleware name to a class name preserving passed parameters.
      *
-     * @param $name
+     * @param  string  $name
      * @return string
      */
     public function resolveMiddlewareClassName($name)
@@ -928,7 +928,7 @@ class Router implements RegistrarContract
      * @param  \Closure|null  $callback
      * @return void
      *
-     * @throws NotFoundHttpException
+     * @throws \Symfony\Component\HttpKernel\Exception\NotFoundHttpException
      */
     public function model($key, $class, Closure $callback = null)
     {
