+++ src/Illuminate/Routing/RoutingServiceProvider.php
@@ -4,8 +4,6 @@ namespace Illuminate\Routing;
 
 use Illuminate\Support\ServiceProvider;
 use Zend\Diactoros\Response as PsrResponse;
-use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\Response;
 use Symfony\Bridge\PsrHttpMessage\Factory\DiactorosFactory;
 
 class RoutingServiceProvider extends ServiceProvider
